\name{deepbayes-package}
\alias{deepbayes-package}
\alias{deepbayes}
\docType{package}
\title{
\packageTitle{deepbayes}
}
\description{
\packageDescription{deepbayes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deepbayes}
\packageIndices{deepbayes}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{deepbayes}

Maintainer: \packageMaintainer{deepbayes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
